Improve Idlib
general:
- Fix unit test excecution
- Make filenames all lowercase

type:
- Fix "scale<uint64_traits>"
- Fix "scale<clamped_double_traits>"

utility:
- Fix "to_string"
- Fix doc comments

crtp:
- Add "assignment_expr"

language:
- Add "category_element"

file system:
- Add "file_descriptor_(windows|osx|linux)"
- Add "mapped_file_descriptor_(windows|linux)"
- Add "access_mode"
- Add "create_mode"
- Add "error"

+++ src/idlib/type/clamped_double_scale.hpp
@@ -33,7 +33,7 @@ struct scale<clamped_double_traits>
 	using traits = clamped_double_traits;
     traits::underlying_type operator()(const traits::underlying_type& v, float s) const
     {
-        return (*this)(v, s);
+        return (*this)(v, double(s));
     }
 
     traits::underlying_type operator()(const traits::underlying_type& v, double s) const
